requirements.txt
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
git-filter-repo==2.47.0
google-api-core==2.25.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.4
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
